refactor: move row-filter helpers into captest.filters module
Move perc_difference, check_all_perc_diff_comb, abs_diff_from_average,
sensor_filter, filter_irr, and filter_grps from capdata.py into a new
filters module. capdata re-imports the four it still uses internally;
test references to perc_difference/abs_diff_from_average now point at
captest.filters (no back-compat shim).
